requirements.txt
Django
djangorestframework
mysqlclient
requests
PyQt5

